+++ defender/__init__.py
@@ -1,3 +1,3 @@
-VERSION = (0, 9, 1)
+VERSION = (0, 9, 4)
 
 __version__ = ".".join((map(str, VERSION)))